+++ tests/php/Integration/Interop/Command/Export/phel-config.php
@@ -3,11 +3,9 @@
 declare(strict_types=1);
 
 use Phel\Config\PhelConfig;
-use Phel\Config\PhelExportConfig;
 
 return new PhelConfig()
-    ->setSrcDirs([__DIR__ . '/src'])
-    ->setExportConfig(new PhelExportConfig()
-        ->setFromDirectories(['src'])
-        ->setNamespacePrefix('PhelTest\Integration\Interop\Command\Export\PhelGenerated')
-        ->setTargetDirectory(__DIR__ . '/PhelGenerated'));
+    ->withSrcDirs([__DIR__ . '/src'])
+    ->withExportFromDirectories(['src'])
+    ->withExportNamespacePrefix('PhelTest\Integration\Interop\Command\Export\PhelGenerated')
+    ->withExportTargetDirectory(__DIR__ . '/PhelGenerated');