+++ tests/php/Integration/Run/Command/Test/TestCommandProjectSuccess/config/phel-config.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 use Phel\Config\PhelConfig;
 
 return new PhelConfig()
-    ->setSrcDirs(['../../../../../../../src/phel/'])
-    ->setTestDirs(['Fixtures'])
-    ->setVendorDir('')
-    ->setErrorLogFile('data/error.log')
-;
+    ->withSrcDirs(['../../../../../../../src/phel/'])
+    ->withTestDirs(['Fixtures'])
+    ->withVendorDir('')
+    ->withErrorLogFile('data/error.log');